Set up PHPDocumentor
This adds the configuration for running [PHPDocumentor](https://github.com/phpDocumentor/phpDocumentor) for PHPCSUtils.

The documentation can be generated with PHPDocumentor 3.x by running the `phpdoc` command from the root of the project.

PHPDocumentor has _not_ been added to the project dependencies as it is only needed intermittely - once for each release - when generating the documentation and doing so with a phar install is recommended by the developers of PHPDocumentor.

The documentation will be published as a [GitHub Pages](https://pages.github.com/) website at the webaddress https://phpcsutils.com/ .