PHPCSUtils: add the ruleset file
This will allow PHPCS to recognize PHPCSUtils as if it were an external PHPCS standard and handle the autoloading of all PHPCSUtils files automatically when using PHPCS 3+.

External standards using PHPCSUtils don't even need to add `<rule ref="PHPCSUtils"/>` to their own ruleset.
All that is needed is for the PHPCSUtils standard to be in the PHPCS `installed_paths` which - for Composer based installs - will automatically be taken care of by the DealerDirect Composer PHPCS plugin which is a required dependency for this package.