:sparkles: New `TestUtils\UtilityMethodTestCase`
This `PHPCSUtils\TestUtils\UtilityMethodTestCase` can be used to test arbitrary utility methods written for PHPCS 2.x as well as PHPCS 3.x, using PHPUnit 4.5 - 8.x.

Aside from taking care of setting up PHPCS and tokenizing the test case file, the class offers two utility functions:
* `getTargetToken()` - to get the token pointer for a target token based on a specific comment found on the/a line before. Returns the int stack pointer to the target.
    Note: the test delimiter comment MUST start with "/* test" to allow this function to distinguish between comments used *in* a test and test delimiters.
* `expectPhpcsException()` - Helper method to tell PHPUnit to expect a PHPCS Exception in a PHPUnit cross-version compatible manner.

If you choose to use this feature and need to use PHPUnit < 5.4 (i.e. if your standard still supports PHP < 5.6), the PHPCSUtils `phpcsutils-autoload.php` file **MUST** be included in your PHPUnit bootstrap file to ensure that the necessary class aliases for PHPUnit are in place.
Similarly, the autoload file **must** be included if you support PHPCS < 3.

By default, the class will look for a test case file in the same directory and with the same file name as the concrete test class, though with an `inc` file extension.

The file extension can be changed by setting the `$fileExtension` property in the concrete class.

To diverge from the default, the path to the test case file can also be provided via the `$caseFile` property, like so:
```php
    /**
     * Initialize & tokenize {@see \PHP_CodeSniffer\Files\File} with code from the test case file.
     *
     * Overloaded to change the `$caseFile`.
     *
     * @beforeClass
     *
     * @return void
     */
    public static function setUpTestFile()
    {
        self::$caseFile = __DIR__ . '/pathToFile.inc';
        parent::setUpTestFile();
    }
```

Includes unit tests for this test utility.