:sparkles: New BackCompat\BCTokens class
PHPCS provides a number of static token arrays in the `PHP_CodeSniffer\Util\Tokens` class.
Some of these token arrays will not be available in older PHPCS versions.
Some will not contain the same set of tokens across PHPCS versions.

The `PHPCSUtils\BackCompat\BCTokens` class is a compatibility layer to allow for retrieving these token arrays with a consistent token content across PHPCS versions.
The one caveat is that the token constants do need to be available.

The class is named `BCTokens` instead of `Tokens` on purpose to more easily allow sniffs to use both the PHPCS native Tokens class as well as this class without always having to alias one of the two.

Recommended usage:
Only use the methods in this class when needed. I.e. when your sniff unit tests indicate a PHPCS cross-version compatibility issue related to inconsistent token arrays.

All PHPCS token arrays are supported, though only a limited number of them are different across PHPCS versions.

The names of the PHPCS native token arrays translate one-on-one to the methods in this class:
- `PHP_CodeSniffer\Util\Tokens::$emptyTokens` => `PHPCSUtils\BackCompat\BCTokens::emptyTokens()`
- `PHP_CodeSniffer\Util\Tokens::$operators`     => `PHPCSUtils\BackCompat\BCTokens::operators()`
    ... etc

The order of the tokens in the arrays may differ between the PHPCS native token arrays and the token arrays returned by this class.

Also note: While `T_LIST` and `T_ANON_CLASS` will be included in the return value for the `parenthesisOpeners()` method, the associated parentheses will not have the `'parenthesis_owner'` index set until PHPCS 3.5.0.

Methods:
* `assignmentTokens()`
* `comparisonTokens()`
* `arithmeticTokens()`
* `operators()`
* `parenthesisOpeners()`
* `phpcsCommentTokens()`
* `textStringTokens()`
* `functionNameTokens()`
* `ooScopeTokens()`
* `__callStatic()` - to retrieve the other token arrays which haven't had any relevant changes between PHPCS 2.6.0 and current.

Includes unit tests for all token arrays.
The unit tests contain a hard-coded copy of the PHPCS native token arrays as they are now.
This means that they will also serve as an early warning system for future changes to these arrays in PHPCS itself.