:sparkles: New Utils\Conditions class
New `PHPCSUtils\Utils\Conditions` class which will hold improved versions of the `File::hasCondition()` and the `File::getCondition()` methods as well as additional methods related to examining a `$tokens[$stackPtr]['conditions']` array.

For the initial functions in this class, the unit tests for the `PHPCSUtils\BackCompat\BCFile::getCondition()` and the `PHPCSUtils\BackCompat\BCFile::hasCondition()` functions are re-used to ensure that the improved functions in this class stay compatible with the original functions.