Conditions::getCondition(): refactor the method to be more versatile
This refactors the `PHPCSUtils\Utils\Conditions::getCondition()` method to be more versatile with an eye on additional methods to be added later.

It also removes duplicate code and will allow the new methods to be added without code duplication.

Includes additional unit tests covering the new functionality added to the `Conditions::getCondition()` method.