BCFile::getDeclarationName(): exclude JS ES6 method support for PHPCS < 3.0.0 [2]
In PHPCS 2.x, for JS files, ES6 method were not tokenized as `T_FUNCTION`.

As the `class` condition is not available either in PHPCS 2.x, it will be hard to back-fill this reliably.

For now, I'm making the choice to explicitly **NOT** backfill support for JS ES6 methods in PHPCS < 3.0.0 for the `BCFile::getDeclarationName()` method.

If enough support requests are opened about it, I may reconsider.

Ref: squizlabs/PHP_CodeSniffer 1251