:sparkles: New Utils\ObjectDeclarations class
New `PHPCSUtils\Utils\ObjectDeclarations` class which will hold improved versions of the `File::getDeclarationName()`, `File::getClassProperties()`, `File::findExtendedClassName()` and the `File::findImplementedInterfaceNames()` methods as well as additional methods related to examining object declaration statements.

For the initial functions in this class, the unit tests for the `PHPCSUtils\BackCompat\BCFile` versions of these functions are re-used to ensure that the improved functions in this class stay compatible with the original functions.

To this end, the pre-existing unit test classes for the `findExtendedClassName()` and `findImplementedInterfaceNames()` methods have been adjusted.