ObjectDeclarations::getName(): improved handling of parse errors
While certain names may be invalid, such as construct names starting with a number, for the purpose of alerting users to these invalid names, the `getName()` method should be able to pick up on them.

The `getName()` method has now been adjusted to allow for such names.

The way this has been implemented also improves the parse error resilience for the `getName()` method.

The `getName()` method could previously inadvertently misidentify the name of a structure in the case of parse errors/live coding as it looked for the first `T_STRING` after the keyword without limiting it to the declaration statement. This has now also been fixed.

The method will now return `null`, both when passed  an anonymous class/function as well as in case of a parse error.

Included additional unit tests.