ObjectDeclarations::findExtendedClassName/findImplementedInterfaceNames(): refactor
As the logic needed for these two methods, as well as for the (upcoming) `findImplementedInterfaces()` method, is basically the same, I've extracted the logic out to a private `ObjectDeclarations::findNames()` method which is now called by both methods to DRY out the code (remove code duplication).

This refactor includes some minor hardening of the methods for the handling of parse errors.

Includes additional unit test covering the parse error hardening.
Includes introducing two new properties to the `PHPCSUtils\Tokens\Collections` class.