:sparkles: New Utils\FunctionDeclarations class
New `PHPCSUtils\Utils\FunctionDeclarations` class which will hold improved versions of the `File::getMethodProperties()` and the `File::getMethodParameters()` methods as well as additional methods related to examining function declaration statements.

Take note of the dropping of `Method` from the method names in the `FunctionDeclarations` class.
As these methods also work for global functions as well as closures, having `Method` in the name doesn't cover the actual functionality of the methods.

For the initial functions in this class, the unit tests for the `PHPCSUtils\BackCompat\BCFile` versions of these functions are re-used to ensure that the improved functions in this class stay compatible with the original functions.

This commit also adds a `getName()` method which is an alias of the `PHPCSUtils\Utils\ObjectDeclarations::getName()` method.