FunctionDeclarations::getProperties(): add 'return_type_end_token' to the array
The `getParameters()` method already returns stack pointers to both the "start" and "end" of a parameter type declaration.

The `getProperties()` method only returned the stack pointer to the _start_ of the return type declaration.

For feature completeness, a new `return_type_end_token` array index has been added to the array which is returned by this method.
The value will be `false` if there is no return type declaration; or the integer stack pointer to the last token of the return type declaration if there is.

Includes a new unit test for this feature, as well as some adjustments to the existing unit tests to allow for this feature.