FunctionDeclarations::getParameters(): code simplification and efficiency fix
Once the equal sign has been found, we know that everything else until the end of the parameter is part of the default value.

So instead of looping through each token individually after the equal sign, go straight to the end of the parameter.

This allows for greatly simplifying the code within various `case`s in the `switch` as well as for merging a lot of the `case`s together.