:sparkles: New Utils\Variables class
New `PHPCSUtils\Utils\Variables` class which will hold an improved version of the `File::getMemberProperties()` method as well as additional methods related to examining variable tokens.

For the initial function in this class, the unit tests for the `PHPCSUtils\BackCompat\BCFile` version of this function are re-used to ensure that the improved function in this class stays compatible with the original function.