Variables::getMemberProperties(): implement use of Scopes class and remove parse error warning
The parse error warning is the purpose of this method, so doesn't belong here.

As an "interface property" is not a valid property, just throw the `$stackPtr is not a class member var` exception instead.

Includes a separate set of unit tests for the changes.

These additional tests would fail when using the BCFile version of the method.