Namespaces::findNamespacePtr(): fix compatibility with PHPCS < 3.3.0
This method runs into a bug in PHPCS where curlies for variable variables received an incorrect and irrelevant scope condition in PHPCS < 3.3.0.

Fixed by just ignoring it if a namespace condition is assigned which is not a namespace declaration.

Note: the fact that the `$prev` pointer is still set to the incorrect "condition" is not an issue as it just means that no other relevant namespace token was found between the two, so skipping ahead is fine.

Ref:
* squizlabs/PHP_CodeSniffer 1882