:sparkles: New Utils\Operators class
New `PHPCSUtils\Utils\Operators` class which will initially hold an improved version of the `File::isReference()` method as well as, at a later stage, additional methods related to examining operator tokens.

For the initial function in this class, the unit tests for the `PHPCSUtils\BackCompat\BCFile::isReference()` function are re-used to ensure that the improved function in this class stays compatible with the original function.