Lists::isShortList(): fix compatibility with PHPCS < 2.8.0 [3]
PHPCS < 2.8.0 contained a bug where a square open bracket `[` directly following a closing curly would always be tokenized as `T_OPEN_SQUARE_BRACKET`, presuming this was array dereferencing/assignment on a variable variable.

If the closing curly however belonged to a control structure, this was incorrect and the token should have been changed to `T_OPEN_SHORT_ARRAY`.

This commit adds work-arounds to the utility method to handle the situation correctly.

Includes unit tests.

Ref: squizlabs/PHP_CodeSniffer 1284