Arrays::isShortArray(): fix compatibility with PHPCS 2.8.0 - 3.2.3 [1]
PHPCS 2.8.0 - 3.2.3 contained a bug where a square open bracket `[` as the first code content after the PHP open tag in a file would be tokenized as `T_OPEN_SQUARE_BRACKET` not `T_OPEN_SHORT_ARRAY`.

This is an edge-case bug as it can only occur when following the very first PHP open tag in a file and only if there is nothing before that open tag.

Either way, this commit adds work-arounds to the utility method to handle the situation correctly.

Includes adding two convenience token arrays for working with arrays in older PHPCS versions to the `PHPCSUtils\Tokens\Collections` class.

Ref: squizlabs/PHP_CodeSniffer 1971