Arrays::isShortArray(): fix compatibility with PHPCS < 2.9.0 [2]
Work around a bug in the tokenizer of PHPCS < 2.9.0 where array dereferencing of short arrays and string literals would be incorrectly tokenized as short array.
I.e. the square brackets in `'PHP'[0]` would be tokenized as short array.

This commit adds work-arounds to the utility method to handle the situation correctly.

Ref: squizlabs/PHP_CodeSniffer 1381