Arrays::isShortArray(): fix compatibility with PHPCS < 2.8.0 [4]
Work around a bug in the tokenizer of PHPCS < 2.8.0 where a `[` would be tokenized as T_OPEN_SQUARE_BRACKET instead of T_OPEN_SHORT_ARRAY if it was preceded by a close curly of a control structure.

This commit adds work-arounds to the utility method to handle the situation correctly.

Ref: squizlabs/PHP_CodeSniffer 1284