:sparkles: New Utils\TextStrings class
This introduces two new methods:
* `getCompleteTextString()` - to retrieve the complete text string content of a - potentially multi-line - text string. Returns the text string.
* `stripQuotes()` - to strip the text delimiter quotes off an arbitrary text string. Returns the resulting text string.

These methods have been battle-tested in various forms in the past year or two in the PHPCompatibility and the WordPressCS standards.

This commit includes a new convenience token array for working with text strings to the `PHPCSUtils\Tokens\Collections` class.

Includes dedicated unit tests for each method.