:sparkles: New `Abstracts\AbstractArrayDeclarationSniff`
New abstract sniff to examine array declarations.

This abstract sets up a number of helpful properties for the array and then calls the following methods from the `processArray()` method:
* `processOpenClose()` to examine the array opener and closer.
    This method will be called independently of whether the array has items or not.
* And then for each individual array item:
    - `processKey()` and `processArrow()` for array items with an index key; or `processNoKey()` for array items without one.
    - `processValue()`
    - `processComma()` (this method _may_ not be called for the last array item if it is not followed by a comma.

Aside from the `process()` method which sets up the properties, all other methods can be overloaded and/or ignored in concrete sniff implementations.
Returning `true` from any of the methods will short-circuit the loop and exit the sniff early.

The properties available to the methods in the child class are:
* `$stackPtr` containing the stack pointer to the array keyword/short array open bracket.
* `$tokens` containing the token stack of the current file.
* `$arrayOpener` containing the stack pointer to the array open parenthesis/short array open bracket.
* `$arrayCloser` containing the stack pointer to the array closer parenthesis/short array close bracket.
* `$arrayItems` containing a multi-dimentional array with information on each array item.
* `$itemCount` containing the number of items found in the array.
* `$singleLine` containing whether or not the array single line or multi-line.

Includes extensive unit tests for the abstract sniff behaviour.