Abstracts\AbstractArrayDeclarationSniff: new `getActualArrayKey()` method
This is a helper method which can optionally be used from the `processKey()` method to retrieve the actual array key value based on the array key rules used in PHP.

* Array keys can only be integers or strings.
* Purely numeric (integer) string keys will automatically be cast to integer.

Dynamic array keys using variables, constants or function calls to set the key are ignored.

The method will return the integer or string array key or void for when the array key could not be determined.

Includes dedicated unit tests.