BCFile::getCondition()/Conditions::getCondition(): sync with PHPCS 3.5.4 / new $first parameter
PHPCS 3.5.4 introduces a new `$first` parameter to the `File::getCondition()` function.

The PHPCSUtils `Conditions::getCondition()` already had a parameter for the same - `$reverse`, though the boolean values used were also the reverse of how the new `$first` parameter is implemented.

As PHPCSUtils is still in `alpha`, I'm making the executive decision to rename the parameter for the `Conditions::getCondition()`  function to `$first`, including reversing the meaning of the boolean values, to stay in line with PHPCS itself.

Includes adjusting any function calls in the existing codebase for the reversal of the meaning of the boolean.
Includes moving the unit tests covering this parameter, which were already in place in the test file for the `Conditions::getCondition()` function, to the test file for the `BCFile::getCondition()` function.
The actual tests will be run for both functions now.

Refs:
* https://github.com/squizlabs/PHP_CodeSniffer/commit/77c5fa2cd0b2ceca49d1610f8a30b857fe0f0fb9.