FunctionDeclaration::getProperties(): document bug fix - 'has_body' = `false` for unfinished arrow functions
The `fn` for a potential arrow function will be tokenized as `T_FN`, even when there are no parenthesis and there is no function body, like during live coding or in case of a parse error.

In that case, IMO, the `has_body` index key should be set to `false`.

This is in line with the behaviour of the `FunctionDeclaration::getProperties()` function for other function constructs (and in contrast to the behaviour of the `BCFile::getMethodProperties()` function).

Also see: PHPCSStandards/PHPCSUtils@198371590e10d7064ae2e19715285a1ed6201896