Arrays::getDoubleArrowPtr(): handle arrow functions
In PHPCS 3.5.3+, the double arrow for arrow functions is tokenized as `T_FN_ARROW` and will not confuse this utility method anymore.

However, for PHPCS < 3.5.3 in combination with PHP 7.4, the `fn` keyword will be tokenized as `T_FN`, however, the double arrow will still tokenize as `T_DOUBLE_ARROW`.
And for PHPCS < 3.5.3 in combination with PHP 7.4, the `fn` keyword will be tokenized as `T_STRING` and the double arrow will tokenize as `T_DOUBLE_ARROW`.

Both of these would cause incorrect results for this function.

As far as I can see, arrow functions are not usable in an array key, so basically, we know that there will be no (array) arrow as soon as we encounter a `T_FN` token of a `T_STRING` token which is an arrow function.

Includes unit tests.

Related: squizlabs/PHP_CodeSniffer 2703

Co-authored-by: Michał Bundyra <contact@webimpress.com>