FunctionDeclarations::isArrowFunction()/getArrowFunctionOpenClose(): sync with PHPCS 3.5.5
This brings the `FunctionDeclarations::isArrowFunction()` and `FunctionDeclarations::getArrowFunctionOpenClose()` functions in line with the changes in PHPCS which will be released in PHPCS 3.5.5.

Note: the return value for the `FunctionDeclarations::getArrowFunctionOpenClose()` function has changed from `array` to `array|false`.
Previously, the function may return an empty array or an array with only some of the indexes set to `false`.
Now it will return `false` when this is not an arrow function or an array with all the indexes set to the relevant stack pointers when it _is_ an arrow function.

This should make the function return value easier to work with.

Includes extensive additional unit tests, as well as some changed unit tests, similar to the additional unit tests as pulled to PHPCS itself for these changes.

Related upstream commits:
* squizlabs/PHP_CodeSniffer@2e3a01035f267bd299cf8ef6fed4b1f19cc2f1dc
* squizlabs/PHP_CodeSniffer@ea810a235ab6865f2355d97b221c068d65820818
* squizlabs/PHP_CodeSniffer@2adcb64a2de94d3993a97e285314c923d6fe7db1
* squizlabs/PHP_CodeSniffer@3a62dd4aa8429821b3f5a9b88852f0e1b854fccb
* squizlabs/PHP_CodeSniffer@6070b6083e0ab6cd3de43dea69ec820884836aac

Related upstream issues/PRs:
* squizlabs/PHP_CodeSniffer 2523
* squizlabs/PHP_CodeSniffer 2859
* squizlabs/PHP_CodeSniffer 2860
* squizlabs/PHP_CodeSniffer 2863