Utils\Numbers: replace `$unsupportedPHPCSVersions` with `UNSUPPORTED_PHPCS_VERSION`
As the fix for the broken backfill in PHPCS 3.5.3 was released in PHPCS 3.5.4, there is only one PHPCS version which is unsupported.

This negates the need for an array, making it possible to use a class constant instead which should be less risky as it cannot be overloaded.

Includes adjusting all places in the code base which were using the property to use the constant instead.