Travis: test on nightly
Up to now, the DealerDirect plugin was a blocker for testing the standard against PHP `nightly`/ PHP 8 as it would refuse to install.

With the new version having been tagged, this issue has been removed, so we can now run the unit tests against PHP 8 to get early warning about cross-version compatibility issues which need fixing.

We still need to use `--ignore-platform-reqs` for PHPUnit as PHPUnit 9 does not (yet) allow for PHP 8, though this is expected to be sorted in v9.3 (expected August).

The builds against `nightly` remain in `allow_failures` for now.

Refs:
* https://github.com/Dealerdirect/phpcodesniffer-composer-installer/releases/tag/v0.7.0