BCFile/ObjectDeclarations::getDeclarationName(): compatibility with PHPCS 4.x
PHPCS removes support for the JS and CSS tokenizers, which includes removing the `File::$tokenizerType` property.

This commit makes it so the `BCFile::getDeclarationName()` and the `ObjectDeclarations::getDeclarationName()` methods won't throw an error for the property not existing.

And just to be clear: while PHPCSUtils in combination with PHPCS 2.x and 3.x _may_ incidentally support JS/CSS, in combination with PHPCS 4.x, JS/CSS will explicitly not be supported.

Ref: https://github.com/squizlabs/PHP_CodeSniffer/commit/ea52e7b48165edbcb96b10a0a5de04567e4d73f8