Lists::getAssignments(): improve the return value
:warning: **BREAKING CHANGE** :warning:

In the initial commit, the `Lists::getAssignments()` method would have three different arrays as potential output.
1. An array with just the `raw` and `is_empty` keys for an empty list item.
2. An array with the above + detailed information about the assignments for non-keyed list items.
3. An array with the above + detail information about the keys found for keyed list items.

This meant that - aside from for the `raw` and `is_empty` keys -, a dev looping over the list items would always need to do an `isset()` check for each array index before the value could be compared or used.

That made the method return value fiddly to work with.

In this commit, the above three different arrays is now reduced to two different arrays (type 2 and 3) with either an empty string or `false` as the default value for all keys.

Additionally:
* The `nested_list` index key has been renamed to `is_nested_list` to make it clearer that this will be a boolean value and not the detailed information about the nested list.
* And while this will rarely be relevant for sniffs implementing this method, the order of the array indexes in the return array has also been adjusted.

Includes updated unit tests.