Namespaces::getType(): improve detection of namespace keyword as operator
When the `namespace` keyword is used in the global namespace, it is less easy to determine with 100% certainty whether it used for a declaration or as an operator, especially when taking into account potential coding errors, such as the use of reserved keywords in the namespace name and/or a leading backslash used in a namespace declaration.

This PR hardens the code and will more often correctly detect whether the namespace keyword is used as an operator when used in the global namespace.

Includes unit tests.