UseStatements::getType(): work around a PHPCS 2.x issue
In PHPCS 2.x, the `case` and `default` keywords are assigned as conditions, even when they are not within a `switch` statement.

The code in question is a parse error, but it also prevents the `getType()` method from correctly resolving the type of a `use` statement, while it is relatively simple to work-around the issue.

So, adding the work-around.

Includes unit tests.