Collections: deprecate $propertyTypeTokens in favour of propertyTypeTokens()
Deprecate the `$propertyTypeTokens` property in favour of a new `Collections::propertyTypeTokens()` method.

PHPCS is expected to introduce a new `T_TYPE_UNION` token. As that token would not always be available, a property can't handle this.

In anticipation of the introduction of the new token, I'm deprecating the `Collections::$propertyTypeTokens` property straight away to prevent BC breaks at a later point.

Includes adding unit tests for the new method.