Collections: deprecate $returnTypeTokens in favour of returnTypeTokens()
Deprecate the `$returnTypeTokens` property in favour of a new `Collections::returnTypeTokens()` method.

PHPCS is expected to introduce a new `T_TYPE_UNION` token. As that token would not always be available, a property can't handle this.

In anticipation of the introduction of the new token, I'm deprecating the `Collections::$returnTypeTokens` property straight away to prevent BC breaks at a later point.

Includes adding unit tests for the new method.