BCFile::getMethodParameters(): sync with PHPCS / union type support
PR 168 added support for union types to the `FunctionDeclarations::getParameters()` method.

Upstream PR 3032 added the same to the PHPCS native `File::getMethodParameters()` method.

As that PR has now been merged, this commit syncs the upstream changes into the `BCFile::getMethodParameters()` method, moves the unit tests to the BCFile test class and makes a minor change to the `FunctionDeclarations::getParameters()` method to support the changes as have been merged upstream.