FunctionDeclarations::getParameters(): add support for PHP 8 constructor property promotion
This commit adds support for constructor property promotion to the `FunctionDeclarations::getParameters()` method.

This change introduces two new keys - `property_visibility` and `visibility_token` - to the return array which will only be included if constructor property promotion is detected.

The method does not check whether property promotion is _valid_ in the function/method in which it is used. That is not the concern of this method.

Changes along the same lines as included in this PR will also be pulled upstream to PHPCS itself. Once the upstream PR has been merged, those changes will also be applied to the `BCFile::getMethodParameters()` method.

Includes unit tests.

Ref: https://wiki.php.net/rfc/constructor_promotion