BCFile::getMethodProperties(): sync with PHPCS / 'return_type_end_token' index
The `return_type_end_token` index as a part of the return value for `FunctionDeclarations::getProperties()` has existed in PHPCSUtils since PR 33 / release `1.0.0-alpha1`.

Upstream PR 3153 has now added this same key to the `File::getMethodProperties()` method.

As that PR has now been merged, this commit syncs the upstream changes in and updates the unit tests to match.