Travis: require PHPUnit 9.3 for PHP 8
PHPUnit 9.3 is not yet released (expected end of August), but is slated to add official PHP 8 support to PHPUnit.

As some tests are failing due to the new reserved keyword `match` being used in PHPUnit, I'm (temporarily) upping the PHPUnit requirement for PHP 8 explicitly to PHPUnit 9.3 (dev).

Once PHPUnit 9.3 has been released, this condition can be removed in favour of changing the requirements in the `composer.json` file.

Note: for now the `--ignore-platform-reqs` is still needed as the underlying dependencies may still not have tagged a version which officially supports PHP 8 and this would cause a requirement conflict.

Ref: https://github.com/sebastianbergmann/phpunit/issues/4373