Collections: deprecate $objectOperators in favour of objectOperators()
Deprecate the `$objectOperators` property in favour of a new `Collections::objectOperators()` method.

PHP 8.0 introduces a new `T_NULLSAFE_OBJECT_OPERATOR` token. As that token will not always be available, a property can't handle this.

In anticipation of the introduction of the new token, I'm deprecating the `Collections::$objectOperators` property straight away to prevent BC breaks at a later point in time.

Includes adding unit tests for the new method.