Tokens\Collections: add new `objectOperatorsBC() method
Sister-method to the `objectOperators()` method, which returns the tokens which can be encountered as object operators in PHP < 8.0 in combination with a PHPCS version in which the token was not backfilled yet.

The backfill for PHPCS has been pulled and is expected to be included in PHPCS 3.5.6 or 3.6.0, but as it stands, is not available in yet in a released version.

Includes unit test.

Related:
* squizlabs/PHP_CodeSniffer 3046