FunctionDeclarations::getProperties(): add tests verifying support for namespace operator in types
The commit which added support for the namespace keyword used as an operator to the `Collections::returnTypeTokens[BC]()` methods implicitly added support for this to the `FunctionDeclarations::getProperties()` method.

This commit adds a unit test to safeguard this support.

The same commit also _silently_ added support for the namespace operator to the `BCFile::getMethodProperties()` method. Namespace operators in return types are not supported yet in PHPCS itself and until they are, the fact that the `BCFile::getMethodProperties()` method supports this should be regarded as an artefact and not as official support.