FunctionDeclarations::getParameters(): add support for namespace operator in types
This commit adds support for the namespace keyword used as an operator to the `FunctionDeclarations::getParameters()` method.

Changes along the same lines as included in this PR will also be pulled upstream to PHPCS itself. Once the upstream PR has been merged, those changes will also be applied to the `BCFile::getMethodParameters()` method.

Includes unit test.