Variables::getMemberProperties(): add tests verifying support for namespace operator in types
The commit which added support for the namespace keyword used as an operator to the `Collections::propertyTypeTokens[BC]()` methods implicitly added support for this to the `Variables::getMemberProperties()` method.

This commit adds a unit test to safeguard this support.

The same commit also _silently_ added support for the namespace operator to the `BCFile::getMemberProperties()` method. Namespace operators in property types are not supported yet in PHPCS itself and until they are, the fact that the `BCFile::getMemberProperties()` method supports this should be regarded as an artefact and not as official support.