PHPCS 4.x | Operators::isReference(): handle closure use as parenthesis owner
`T_USE` tokens used for closure use statements are parentheses owners as of PHPCS 4.x.

This makes a minor code adjustment to handle this in the `Operators::isReference()` method.

The previous code would already handle it correctly, but this adds some forward-compatibility code for when support for PHPCS < 4 would be dropped.

Refs:
* squizlabs/PHP_CodeSniffer 2593
* https://github.com/squizlabs/PHP_CodeSniffer/commit/08824f327ce5e20d7528c5e838dd7fade3dcd11c#
* squizlabs/PHP_CodeSniffer 3104
* squizlabs/PHP_CodeSniffer xxxx