BCFile::isReference(): sync with PHPCS / arrow function params passed by ref
PR 192 added support for recognizing the reference operator when used for parameters passed by reference in arrow functions to the `Operators::isReference()` method.

Upstream PR 3103 added the same to the PHPCS native `File::isReference()` method.

As that PR has now been merged and will be included in PHPCS 3.5.7, this commit syncs the upstream changes in, by moving the unit tests from the "Diff" tests to the BCFile tests and annotating that parameters passed by reference for arrow functions are now officially supported in the `BCFile::isReference()` method.