PHPCS 4.x | BCTokens::parenthesisOpeners(): T_USE is now a parenthesis opener
`T_USE` tokens used for closure use statements are parentheses openers as of PHPCS 4.x.

This includes the token having been added to the `Tokens::$parenthesisOpeners` array.

This syncs that change into PHPCSUtils.

The method itself does not need to change as it will automatically pick up on this.
The unit test, however, does need updating.

Refs:
* squizlabs/PHP_CodeSniffer 2593
* https://github.com/squizlabs/PHP_CodeSniffer/commit/08824f327ce5e20d7528c5e838dd7fade3dcd11c#